put debug output behind a flag.
Every time you open a javascript file, it echoes the fact that it added
node_modules to the path, which, when the package is operating as
expected is mostly noise.

This enhances the debug output to not only include the exact
directories which are added to the `exec-path`, but also to only
output when the debug flag, `add-node-modules-path-debug` is turned
on.